chore: rename `raspberry_pi_helper` to `device_helper`
