+++ lib/device_helper.py
