+++ lib/github.py
@@ -13,7 +13,7 @@ from requests import (
 )
 from lib.utils import is_balena_app, is_docker, is_ci, connect_to_redis
 from lib.diagnostics import get_git_branch, get_git_hash, get_git_short_hash
-from lib.raspberry_pi_helper import parse_cpu_info
+from lib.device_helper import parse_cpu_info
 from settings import settings
