+++ lib/media_player.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 import sh
 import vlc
 
-from lib.raspberry_pi_helper import get_device_type
+from lib.device_helper import get_device_type
 from settings import settings
 
 VIDEO_TIMEOUT = 20  # secs
